// src/app/enums/card5-key.enum.ts
export enum Card5KeyEnum {
  Background = 'background',
  Sell = 'sell',
  SellBox = 'sell-box',
  Coin = 'coin',
  Type = 'type',
  Factor = 'factor',
  RiskBox = 'risk-box',
  Pnl = 'pnl',
  Roe = 'roe',
  Size = 'size',
  Margin = 'margin',
  Risk = 'risk',
  EntryPrice = 'entry-price',
  LastPrice = 'last-price',
  LiquidPrice = 'liquid-price'
}